Make test plugin
