+++ gazebo_tutorials/launch/hello.launch
@@ -1,0 +1,7 @@
+<launch>
+  <!-- We resume the logic in empty_world.launch, changing only the name of the world to be launched -->
+  <include file="$(find gazebo_ros)/launch/empty_world.launch">
+    <arg name="world_name" value="$(find gazebo_tutorials)/worlds/hello.world"/>
+    <!-- more default parameters can be changed here -->
+  </include>
+</launch>
