+++ gazebo_tutorials/src/simple_world_plugin.cpp
@@ -1,0 +1,17 @@
+#include <gazebo/gazebo.hh>
+
+namespace gazebo
+{
+    class WorldPluginTutorial : public WorldPlugin
+    {
+    public: WorldPluginTutorial() : WorldPlugin()
+        {
+            printf("Hello World!\n");
+        }
+
+    public: void Load(physics::WorldPtr _world, sdf::ElementPtr _sdf)
+        {
+        }
+    };
+    GZ_REGISTER_WORLD_PLUGIN(WorldPluginTutorial)
+}
