--- march_gazebo_plugins/CMakeLists.txt
@@ -1,11 +1,0 @@
-cmake_minimum_required(VERSION 2.8 FATAL_ERROR)
-
-find_package(gazebo REQUIRED)
-include_directories(${GAZEBO_INCLUDE_DIRS})
-link_directories(${GAZEBO_LIBRARY_DIRS})
-list(APPEND CMAKE_CXX_FLAGS "${GAZEBO_CXX_FLAGS}")
-
-add_library(hello_world SHARED hello_world.cc)
-target_link_libraries(hello_world ${GAZEBO_LIBRARIES})
-
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${GAZEBO_CXX_FLAGS}")
