+++ march_simulation/CMakeLists.txt
@@ -9,6 +9,7 @@ find_package(catkin REQUIRED COMPONENTS
     joint_state_controller
     joint_trajectory_controller
     march_description
+    gazebo_tutorials
     robot_state_publisher
     xacro
 )
