+++ march_simulation/package.xml
@@ -13,6 +13,7 @@
     <depend>gazebo_ros</depend>
     <depend>gazebo_ros_control</depend>
     <depend>march_description</depend>
+    <depend>gazebo_tutorials</depend>
     <depend>robot_state_publisher</depend>
     <depend>joint_state_controller</depend>
     <depend>joint_trajectory_controller</depend>
